Remove `hasDeprecationWarnings` from `watch` function, move it up one level.
