+++ packages/jest-cli/src/__tests__/watch-filename-pattern-mode-test.js
@@ -84,7 +84,6 @@ describe('Watch mode flows', () => {
   let argv;
   let context;
   let config;
-  let hasDeprecationWarnings;
   let stdin;
 
   beforeEach(() => {
@@ -94,13 +93,12 @@ describe('Watch mode flows', () => {
     argv = {};
     context = {};
     config = {};
-    hasDeprecationWarnings = false;
     stdin = new MockStdin();
   });
 
   it('Pressing "P" enters pattern mode', () => {
     config = {rootDir: ''};
-    watch(config, pipe, argv, hasteMap, context, hasDeprecationWarnings, stdin);
+    watch(config, pipe, argv, hasteMap, context, stdin);
 
     // Write a enter pattern mode
     stdin.emit(KEYS.P);
@@ -137,7 +135,7 @@ describe('Watch mode flows', () => {
 
   it('Results in pattern mode get truncated appropriately', () => {
     config = {rootDir: ''};
-    watch(config, pipe, argv, hasteMap, context, hasDeprecationWarnings, stdin);
+    watch(config, pipe, argv, hasteMap, context, stdin);
 
     stdin.emit(KEYS.P);
