+++ packages/jest-cli/src/__tests__/watch-test.js
@@ -41,7 +41,6 @@ describe('Watch mode flows', () => {
   let argv;
   let context;
   let config;
-  let hasDeprecationWarnings;
   let stdin;
 
   beforeEach(() => {
@@ -50,7 +49,6 @@ describe('Watch mode flows', () => {
     argv = {};
     context = {};
     config = {roots: [], testPathIgnorePatterns: [], testRegex: ''};
-    hasDeprecationWarnings = false;
     stdin = new MockStdin();
   });
 
@@ -58,7 +56,7 @@ describe('Watch mode flows', () => {
     argv.testPathPattern = 'test-*';
     config.testPathPattern = 'test-*';
 
-    watch(config, pipe, argv, hasteMap, context, hasDeprecationWarnings, stdin);
+    watch(config, pipe, argv, hasteMap, context, stdin);
 
     expect(runJestMock).toBeCalledWith(
       [context],
@@ -74,7 +72,7 @@ describe('Watch mode flows', () => {
     argv.testNamePattern = 'test-*';
     config.testNamePattern = 'test-*';
 
-    watch(config, pipe, argv, hasteMap, context, hasDeprecationWarnings, stdin);
+    watch(config, pipe, argv, hasteMap, context, stdin);
 
     expect(runJestMock).toBeCalledWith(
       [context],
@@ -87,7 +85,7 @@ describe('Watch mode flows', () => {
   });
 
   it('Runs Jest once by default and shows usage', () => {
-    watch(config, pipe, argv, hasteMap, context, hasDeprecationWarnings, stdin);
+    watch(config, pipe, argv, hasteMap, context, stdin);
     expect(runJestMock).toBeCalledWith(
       [context],
       argv,
@@ -100,7 +98,7 @@ describe('Watch mode flows', () => {
   });
 
   it('Pressing "o" runs test in "only changed files" mode', () => {
-    watch(config, pipe, argv, hasteMap, context, hasDeprecationWarnings, stdin);
+    watch(config, pipe, argv, hasteMap, context, stdin);
     runJestMock.mockReset();
 
     stdin.emit(KEYS.O);
@@ -114,7 +112,7 @@ describe('Watch mode flows', () => {
   });
 
   it('Pressing "a" runs test in "watch all" mode', () => {
-    watch(config, pipe, argv, hasteMap, context, hasDeprecationWarnings, stdin);
+    watch(config, pipe, argv, hasteMap, context, stdin);
     runJestMock.mockReset();
 
     stdin.emit(KEYS.A);
@@ -128,14 +126,14 @@ describe('Watch mode flows', () => {
   });
 
   it('Pressing "ENTER" reruns the tests', () => {
-    watch(config, pipe, argv, hasteMap, context, hasDeprecationWarnings, stdin);
+    watch(config, pipe, argv, hasteMap, context, stdin);
     expect(runJestMock).toHaveBeenCalledTimes(1);
     stdin.emit(KEYS.ENTER);
     expect(runJestMock).toHaveBeenCalledTimes(2);
   });
 
   it('Pressing "u" reruns the tests in "update snapshot" mode', () => {
-    watch(config, pipe, argv, hasteMap, context, hasDeprecationWarnings, stdin);
+    watch(config, pipe, argv, hasteMap, context, stdin);
     runJestMock.mockReset();
 
     stdin.emit(KEYS.U);
