+++ packages/jest-cli/src/watch.js
@@ -34,17 +34,8 @@ const watch = (
   argv: Object,
   hasteMap: HasteMap,
   context: Context,
-  hasDeprecationWarnings?: boolean,
   stdin?: stream$Readable | tty$ReadStream = process.stdin,
 ) => {
-  if (hasDeprecationWarnings) {
-    return handleDeprecatedWarnings(pipe, stdin)
-      .then(() => {
-        watch(config, pipe, argv, hasteMap, context);
-      })
-      .catch(() => process.exit(0));
-  }
-
   setState(argv, argv.watch ? 'watch' : 'watchAll', {
     testNamePattern: argv.testNamePattern,
     testPathPattern: argv.testPathPattern ||
@@ -225,39 +216,6 @@ const watch = (
 
   startRun();
   return Promise.resolve();
-};
-
-const handleDeprecatedWarnings = (
-  pipe: stream$Writable | tty$WriteStream,
-  stdin: stream$Readable | tty$ReadStream = process.stdin,
-) => {
-  return new Promise((resolve, reject) => {
-    if (typeof stdin.setRawMode === 'function') {
-      const messages = [
-        chalk.red('There are deprecation warnings.\n'),
-        chalk.dim(' \u203A Press ') + 'Enter' + chalk.dim(' to continue.'),
-        chalk.dim(' \u203A Press ') + 'Esc' + chalk.dim(' to exit.'),
-      ];
-
-      pipe.write(messages.join('\n'));
-
-      // $FlowFixMe
-      stdin.setRawMode(true);
-      stdin.resume();
-      stdin.setEncoding('hex');
-      stdin.on('data', (key: string) => {
-        if (key === KEYS.ENTER) {
-          resolve();
-        } else if (
-          [KEYS.ESCAPE, KEYS.CONTROL_C, KEYS.CONTROL_D].indexOf(key) !== -1
-        ) {
-          reject();
-        }
-      });
-    } else {
-      resolve();
-    }
-  });
 };
 
 const usage = (argv, snapshotFailure, delimiter = '\n') => {
